Rename folder from estimate_job_count to estimate_filled_posts
