+++ jobs/estimate_ind_cqc_filled_posts.py
@@ -14,13 +14,13 @@ from utils.column_names.ind_cqc_pipeline_columns import (
     PartitionKeys as Keys,
     IndCqcColumns as IndCqc,
 )
-from utils.estimate_job_count.models.care_homes import model_care_homes
-from utils.estimate_job_count.models.primary_service_rolling_average import (
+from utils.estimate_filled_posts.models.care_homes import model_care_homes
+from utils.estimate_filled_posts.models.primary_service_rolling_average import (
     model_primary_service_rolling_average,
 )
-from utils.estimate_job_count.models.extrapolation import model_extrapolation
-from utils.estimate_job_count.models.interpolation import model_interpolation
-from utils.estimate_job_count.models.non_res_with_pir import (
+from utils.estimate_filled_posts.models.extrapolation import model_extrapolation
+from utils.estimate_filled_posts.models.interpolation import model_interpolation
+from utils.estimate_filled_posts.models.non_res_with_pir import (
     model_non_residential_with_pir,
 )
