+++ tests/unit/test_model_primary_service_rolling_average.py
@@ -4,7 +4,7 @@ import pyspark.sql.functions as F
 
 from pyspark.sql import SparkSession
 
-import utils.estimate_job_count.models.primary_service_rolling_average as job
+import utils.estimate_filled_posts.models.primary_service_rolling_average as job
 from tests.test_file_data import ModelPrimaryServiceRollingAverage as Data
 from tests.test_file_schemas import ModelPrimaryServiceRollingAverage as Schemas
 from utils.column_names.ind_cqc_pipeline_columns import (
