+++ utils/diagnostics_utils/diagnostics_meta_data.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from utils.estimate_job_count.column_names import (
+from utils.estimate_filled_posts.column_names import (
     PEOPLE_DIRECTLY_EMPLOYED,
     JOB_COUNT_UNFILTERED,
     JOB_COUNT,
